Table docs split, table docs updated, utilities improved
