+++ src/routes/(inner)/components/tables/data-tables/+page.svelte
@@ -1,0 +1,47 @@
+<script lang="ts">
+	import DocsShell from '$docs/DocsShell/DocsShell.svelte';
+	import { DocsFeature, type DocsShellSettings } from '$docs/DocsShell/types';
+
+	// Components
+	import DataTable from '$lib/components/Table/DataTable.svelte';
+	// Utilities
+	import CodeBlock from '$lib/utilities/CodeBlock/CodeBlock.svelte';
+
+	// @ts-ignore
+	import sveldDataTable from '$lib/components/Table/DataTable.svelte?raw&sveld';
+
+	// Docs Shell
+	const settings: DocsShellSettings = {
+		feature: DocsFeature.Component,
+		name: 'Data Tables',
+		description: 'Feature-rich tables with support for selection, sort, and more.',
+		imports: ['DataTable'],
+		// types: ['XXX'],
+		source: 'components/Table',
+		aria: 'https://www.w3.org/WAI/ARIA/apg/patterns/grid/',
+		components: [{ sveld: sveldDataTable }],
+		keyboard: [
+			['<kbd>Right Arrow</kbd>', 'Moves focus one cell to the right. If focus is on the right-most cell in the row, focus does not move.'],
+			['<kbd>Left Arrow</kbd>', 'Moves focus one cell to the left. If focus is on the left-most cell in the row, focus does not move.'],
+			['<kbd>Down Arrow</kbd>', 'Moves focus one cell down. If focus is on the bottom cell in the column, focus does not move.'],
+			['<kbd>Up Arrow</kbd>', 'Moves focus one cell Up. If focus is on the top cell in the column, focus does not move.'],
+			['<kbd>Home</kbd>', 'Moves focus to the first cell in the row that contains focus.'],
+			['<kbd>End</kbd>', ' Moves focus to the last cell in the row that contains focus.'],
+			['<kbd>Enter</kbd> or <kbd>Space</kbd>', 'Triggers the on:selected event.']
+		]
+	};
+</script>
+
+<DocsShell {settings}>
+	<!-- Slot: Sandbox -->
+	<svelte:fragment slot="sandbox">
+		<p>(examples)</p>
+	</svelte:fragment>
+
+	<!-- Slot: Usage -->
+	<svelte:fragment slot="usage">
+		<section class="space-y-4">
+			<p>(usage)</p>
+		</section>
+	</svelte:fragment>
+</DocsShell>
